docparseai/extractors/image_extractor.py
\"""Image document extractor for DocParseAI.

This module provides functionality to extract text and other content
from image-based documents using OCR technology.
"""

from typing import Dict, List, Optional, Union
from pathlib import Path
import warnings

# Try to import OCR libraries
try:
    import pytesseract
    from PIL import Image
    LIBRARIES_AVAILABLE = True
except ImportError:
    LIBRARIES_AVAILABLE = False
    warnings.warn(
        "OCR processing libraries are not installed. "
        "Install required packages with: pip install -r requirements.txt"
    )


class ImageExtractor:
    """Extractor for image-based documents.
    
    This class provides methods to extract text and other content from images
    using Optical Character Recognition (OCR) technology.
    
    Attributes:
        config (Dict): Configuration options for the image extractor.
    """
    
    def __init__(self, config: Optional[Dict] = None):
        """Initialize a new ImageExtractor instance.
        
        Args:
            config: Optional configuration dictionary for customizing extraction behavior.
        """
        self.config = config or {}
        # TODO: Add support for OCR libraries like Tesseract (via pytesseract)
    
    def extract_text(self, image_path: Union[str, Path]) -> str:
        """Extract text content from an image using OCR.
        
        Args:
            image_path: Path to the image file.
            
        Returns:
            str: The extracted text content.
            
        Raises:
            FileNotFoundError: If the image file doesn't exist.
            ValueError: If the image format is not supported.
        """
        image_path = Path(image_path)
        
        if not image_path.exists():
            raise FileNotFoundError(f"Image file not found: {image_path}")
        
        # Check if the file is a supported image format
        file_ext = image_path.suffix.lower()
        if file_ext not in (".jpg", ".jpeg", ".png", ".tiff", ".bmp"):
            raise ValueError(f"Unsupported image format: {file_ext}")
        
        if not LIBRARIES_AVAILABLE:
            raise RuntimeError("OCR processing libraries are not installed")
        
        # Open the image file
        with Image.open(image_path) as img:
            # Use pytesseract to extract text
            extracted_text = pytesseract.image_to_string(img)
        
        return extracted_text
    
    def extract_tables(self, image_path: Union[str, Path]) -> List[Dict]:
        """Extract tables from an image using OCR and table detection.
        
        Args:
            image_path: Path to the image file.
            
        Returns:
            List[Dict]: List of extracted tables, where each table is represented as a dictionary
                       with keys for headers and data.
        """
        image_path = Path(image_path)
        
        if not image_path.exists():
            raise FileNotFoundError(f"Image file not found: {image_path}")
        
        if not LIBRARIES_AVAILABLE:
            raise RuntimeError("OCR processing libraries are not installed")
        
        # Open the image file
        with Image.open(image_path) as img:
            # Use pytesseract to extract tables
            # This is a placeholder for actual table extraction logic
            extracted_tables = pytesseract.image_to_data(img, output_type=pytesseract.Output.DICT)
        
        # TODO: Implement actual table extraction logic using extracted_tables
        return [{
            "headers": ["Column 1", "Column 2", "Column 3"],
            "data": [
                ["Row 1, Cell 1", "Row 1, Cell 2", "Row 1, Cell 3"],
                ["Row 2, Cell 1", "Row 2, Cell 2", "Row 2, Cell 3"],
            ]
        }]
    
    def extract_form_fields(self, image_path: Union[str, Path]) -> Dict:
        """Extract form fields from an image using OCR and field detection.
        
        Args:
            image_path: Path to the image file.
            
        Returns:
            Dict: Dictionary of extracted form fields, where keys are field names and values are field values.
        """
        image_path = Path(image_path)
        
        if not image_path.exists():
            raise FileNotFoundError(f"Image file not found: {image_path}")
        
        if not LIBRARIES_AVAILABLE:
            raise RuntimeError("OCR processing libraries are not installed")
        
        # Open the image file
        with Image.open(image_path) as img:
            # Use pytesseract to extract form fields
            # This is a placeholder for actual form field extraction logic
            extracted_fields = pytesseract.image_to_data(img, output_type=pytesseract.Output.DICT)
        
        # TODO: Implement actual form field extraction logic using extracted_fields
        return {
            "field1": "Value 1",
            "field2": "Value 2",
        }