requirements.txt
# DocParseAI Core Requirements

# PDF Processing
PyPDF2>=2.0.0
pdfplumber>=0.7.0
tabula-py>=2.3.0
camelot-py>=0.10.1
pdf2image>=1.16.0

# Image Processing
pillow>=9.0.0
pytesseract>=0.3.9

# Document Processing
python-docx>=0.8.11

# Utilities
numpy>=1.22.0
pandas>=1.4.0